requirements.txt
fastapi==0.116.1
langchain_community==0.4.1
langchain_core==1.0.4
langchain_groq==1.0.0
pandas==2.3.3
dotenv==0.9.9
pydantic==2.12.4
python-dotenv==1.1.1
yfinance==0.2.65
regex==2025.7.34
uvicorn==0.35.0
pinecone==7.3.0
langchain-pinecone==0.2.13
sentence-transformers==5.0.0
tabulate==0.9.0
